@@ -11,18 +11,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.15",
+    "@angular/animations": "^13.1.1",
     "@angular/cdk": "^10.2.7",
-    "@angular/common": "~12.2.15",
-    "@angular/compiler": "~12.2.15",
-    "@angular/core": "~12.2.15",
-    "@angular/forms": "~12.2.15",
-    "@angular/localize": "~12.2.15",
+    "@angular/common": "~13.1.1",
+    "@angular/compiler": "~13.1.1",
+    "@angular/core": "~13.1.1",
+    "@angular/forms": "~13.1.1",
+    "@angular/localize": "~13.1.1",
     "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "~12.2.15",
-    "@angular/platform-browser-dynamic": "~12.2.15",
-    "@angular/router": "~12.2.15",
+    "@angular/platform-browser": "~13.1.1",
+    "@angular/platform-browser-dynamic": "~13.1.1",
+    "@angular/router": "~13.1.1",
     "@ng-bootstrap/ng-bootstrap": "^8.0.4",
+    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/http-loader": "^7.0.0",
     "bootstrap": "^4.5.0",
     "chart.js": "^2.9.4",
     "jquery": "^3.6.0",
@@ -33,12 +35,12 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.14",
-    "@angular/cli": "~12.2.14",
-    "@angular/compiler-cli": "~12.2.15",
-    "@types/node": "^12.11.1",
+    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular/cli": "~13.1.2",
+    "@angular/compiler-cli": "~13.1.1",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
@@ -50,6 +52,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.5.4"
   }
 }
