@@ -26,56 +26,54 @@ import { BeetreportComponent } from '../modules/BEET/beetreport/beetreport.compo
 
 
 @NgModule({
-  declarations: [
-    HeaderComponent,
-    HeaderTwoComponent,
-    FooterComponent,
-    PieChartComponent,
-    BarChartComponent,
-    DialogComponent,
-    ConfirmationDialogComponent,
-    textfieldsdialogComponent,
-    WindowRdialogComponent,
-    OuterwallRadvancedleveldialogComponent,
-    RoofradvancedComponent,
-    FooterTwoComponent,
-    BeetreportpdfComponent,
-  ],
-  imports: [
-    CommonModule,
-    MaterialModule,
-    FormsModule,
-    ReactiveFormsModule,
-    NgbModule,
-    ChartsModule,
-    MatDialogModule,
-    RouterModule
+    declarations: [
+        HeaderComponent,
+        HeaderTwoComponent,
+        FooterComponent,
+        PieChartComponent,
+        BarChartComponent,
+        DialogComponent,
+        ConfirmationDialogComponent,
+        textfieldsdialogComponent,
+        WindowRdialogComponent,
+        OuterwallRadvancedleveldialogComponent,
+        RoofradvancedComponent,
+        FooterTwoComponent,
+        BeetreportpdfComponent,
     ],
-  exports: [
-    HeaderComponent,
-    FooterComponent,
-    CommonModule,
-    MaterialModule,
-    FormsModule,
-    ReactiveFormsModule,
-    NgbModule,
-    ChartsModule,
-    PieChartComponent,
-    BarChartComponent,
-    MatDialogModule,
-    ConfirmationDialogComponent,
-    DialogComponent,
-    textfieldsdialogComponent,
-    WindowRdialogComponent,
-    OuterwallRadvancedleveldialogComponent,
-    RoofradvancedComponent,
-    HeaderTwoComponent,
-    FooterTwoComponent,
-    BeetreportpdfComponent
-  ],
-  
-  // If Service do not have Provider == root then add the below code to use the service by other components.
-  providers: [ ConfirmationDialogService,NgbActiveModal ],
-  entryComponents: [ ConfirmationDialogComponent ],
+    imports: [
+        CommonModule,
+        MaterialModule,
+        FormsModule,
+        ReactiveFormsModule,
+        NgbModule,
+        ChartsModule,
+        MatDialogModule,
+        RouterModule
+    ],
+    exports: [
+        HeaderComponent,
+        FooterComponent,
+        CommonModule,
+        MaterialModule,
+        FormsModule,
+        ReactiveFormsModule,
+        NgbModule,
+        ChartsModule,
+        PieChartComponent,
+        BarChartComponent,
+        MatDialogModule,
+        ConfirmationDialogComponent,
+        DialogComponent,
+        textfieldsdialogComponent,
+        WindowRdialogComponent,
+        OuterwallRadvancedleveldialogComponent,
+        RoofradvancedComponent,
+        HeaderTwoComponent,
+        FooterTwoComponent,
+        BeetreportpdfComponent
+    ],
+    // If Service do not have Provider == root then add the below code to use the service by other components.
+    providers: [ConfirmationDialogService, NgbActiveModal]
 })
 export class SharedModule { }
